Move performance tests to ComposedPerformanceTests
